feat(library): Code refactored
