+++ .gitignore
@@ -1,0 +1,4 @@
+DynDnsUpdater.build
+EagerLinkingTBDs
+Release
+XCBuildData
