+++ DynDnsUpdaterUI/DynDnsUpdaterUIApp.swift
@@ -11,10 +11,10 @@ import AppKit
 
 @main
 struct DynDnsUpdaterUIApp: App {
-    
+
     @NSApplicationDelegateAdaptor(AppDelegate.self)
     private var appDelegate
-    
+
     var body: some Scene {
         WindowGroup {
             ContentView()
